requirements.txt
Click
lxml
colorama
pyYAML
wmi
beautifulsoup4==4.9.3
requests==2.24.0